Feat: RealmList 활용하여 찜 탭 segment 분류 및 검색 기능 구현

+++ ToBuy/Liked/LikedItemViewController.swift
@@ -15,8 +15,10 @@ final class LikedItemViewController: BaseViewController {
     
     let realm = try! Realm()
     let repository = LikedItemTableRepository()
+    let folderRepository = FolderRepository()
     lazy var liked = repository.fetchAlls()
-    var list: [Folder] = []
+    lazy var list = folderRepository.fetchFolder()
+    var segmentIndex = 0
     
     let calendar = {
         let view = FSCalendar()
@@ -26,11 +28,14 @@ final class LikedItemViewController: BaseViewController {
 
     let segment = {
         let control = UISegmentedControl(items: ["상품", "브랜드", "물욕달력"])
+        control.selectedSegmentIndex = 0
         return control
     }()
     
     let searchBar = UISearchBar()
     lazy var collectionView = UICollectionView(frame: .zero, collectionViewLayout: collectionViewLayout())
+    
+    let tableView = UITableView() 
     
     
     override func viewDidLoad() {
@@ -39,22 +44,32 @@ final class LikedItemViewController: BaseViewController {
         collectionView.delegate = self
         collectionView.dataSource = self
         collectionView.register(LikedItemCollectionViewCell.self, forCellWithReuseIdentifier:   LikedItemCollectionViewCell.identifier)
+        
+        tableView.delegate = self
+        tableView.dataSource = self
+        tableView.register(likedVCBrandTableViewCell.self, forCellReuseIdentifier: likedVCBrandTableViewCell.identifier)
+        
         searchBar.delegate = self
         
-        list = repository.fetchFolder()
+        list = folderRepository.fetchFolder()
         print(repository.detectRealmURL())
-        print(list)
+
     }
     
     override func viewWillAppear(_ animated: Bool) {
         print(#function)
         liked  = repository.fetchAlls()
+        list = folderRepository.fetchFolder()
         collectionView.reloadData()
+        tableView.reloadData()
         configView()
+        
+        
     }
     override func configHierarchy() {
         view.addSubview(searchBar)
         view.addSubview(collectionView)
+        view.addSubview(tableView)
         view.addSubview(segment)
         view.addSubview(calendar)
     }
@@ -76,6 +91,11 @@ final class LikedItemViewController: BaseViewController {
             make.horizontalEdges.bottom.equalTo(view.safeAreaLayoutGuide)
         }
         
+        tableView.snp.makeConstraints { make in
+            make.top.equalTo(segment.snp.bottom).offset(10)
+            make.horizontalEdges.bottom.equalTo(view.safeAreaLayoutGuide)
+        }
+        
         calendar.snp.makeConstraints { make in
             make.top.equalTo(segment.snp.bottom).offset(10)
             make.horizontalEdges.equalTo(view.safeAreaLayoutGuide)
@@ -88,15 +108,12 @@ final class LikedItemViewController: BaseViewController {
         navigationItem.title = "찜 \(liked.count)개 목록"
         hideKeyboardWhenTappedAround()
         segment.addTarget(self, action: #selector(segmentChanged(segment:)), for: .valueChanged)
-        
-        
+        tableView.rowHeight = 300
     }
 }
 
 extension LikedItemViewController {
-    
-    
-    
+     
     func collectionViewLayout() -> UICollectionViewLayout {
         let layout = UICollectionViewFlowLayout()
         let width = UIScreen.main.bounds.width - 40
@@ -110,10 +127,32 @@ extension LikedItemViewController {
     }
     
     @objc func segmentChanged(segment: UISegmentedControl) {
-        
+        print(#function)
+     
         let selectedIndex = segment.selectedSegmentIndex
-        calendar.isHidden = selectedIndex == 0 || selectedIndex == 1
-        collectionView.isHidden = selectedIndex == 2
+        print(selectedIndex)
+        switch selectedIndex {
+           case 0:
+               segmentIndex = 0
+               calendar.isHidden = true
+               tableView.isHidden = true
+                collectionView.isHidden = false
+                collectionView.reloadData()
+           case 1:
+               segmentIndex = 1
+               calendar.isHidden = true
+               collectionView.isHidden = false
+               tableView.isHidden = false
+            tableView.reloadData()
+           case 2:
+               segmentIndex = 2
+                collectionView.isHidden = true
+               tableView.isHidden = true
+               calendar.isHidden = false
+           default:
+               break
+           }
+        viewWillAppear(true)
     }
     
     @objc func likedBtnTapped(_ btn: UIButton){
@@ -121,32 +160,67 @@ extension LikedItemViewController {
         print(#function)
         let index = btn.tag
         let all = repository.fetchAlls()
+        let folder = all[index].main.first
+        guard let folder = folder else {return}
         
-        repository.deleteItem(id: all[index].id)
-        print("Product deleted")
-        
+        if folder.detail.count == 1 {
+            folderRepository.removeFolder(folder)
+        } else if folder.detail.count > 1 {
+            repository.deleteItem(id: all[index].id)
+            print("Product deleted")
+        }
         viewWillAppear(true)
         configView()
     }
 }
 
+extension LikedItemViewController: UITableViewDelegate, UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        print(list)
+        return list.count
+        
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: likedVCBrandTableViewCell.identifier, for: indexPath) as! likedVCBrandTableViewCell
+        
+        cell.configUI(data: list[indexPath.row])
+        
+        cell.collectionView.delegate = self
+        cell.collectionView.dataSource = self
+        cell.collectionView.register(LikedItemCollectionViewCell.self, forCellWithReuseIdentifier: LikedItemCollectionViewCell.identifier)
+        cell.collectionView.tag = indexPath.row
+        cell.collectionView.reloadData()
+        
+        return cell
+    }
 
+}
 extension LikedItemViewController: UICollectionViewDelegate, UICollectionViewDataSource {
     
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        
-        return liked.count
+        if collectionView == self.collectionView {
+            return liked.count
+        } else {
+            return list[collectionView.tag].detail.count
+        }
     }
     
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         let cell = collectionView.dequeueReusableCell(withReuseIdentifier: LikedItemCollectionViewCell.identifier, for: indexPath) as! LikedItemCollectionViewCell
         
         cell.layer.cornerRadius = 10
-        
-        cell.configUI(data: liked[indexPath.item])
         cell.likeBtn.addTarget(self, action: #selector(likedBtnTapped), for: .touchUpInside)
         cell.likeBtn.tag = indexPath.item
-        return cell
+        
+        if collectionView == self.collectionView {
+            cell.configUI(data: liked[indexPath.item])
+        
+            return cell
+        } else {
+            cell.configUI(data: list[collectionView.tag].detail[indexPath.item])
+            return cell
+        }
     }
     
 //    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
@@ -163,11 +237,15 @@ extension LikedItemViewController: UISearchBarDelegate {
     
     func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
         let filter = realm.objects(LikedItemTable.self).where { $0.title.contains(searchText, options: .caseInsensitive) }
-        
+        let folderFilter = realm.objects(Folder.self).where {
+            $0.detail.title.contains(searchText, options: .caseInsensitive)
+        }
         liked = filter
+        list = folderFilter
+        
         collectionView.reloadData()
+        tableView.reloadData()
         
     }
-    
-    
+
 }
